Move the resending method in a separate file that could contain other utilities & improve comment
